refactor: dry things out mroe
